+++ packages/cli/src/commands/link/unlink.js
@@ -9,8 +9,7 @@
 
 import {flatMap, values, difference} from 'lodash';
 import type {ConfigT} from '../../tools/config/types.flow';
-import dedent from 'dedent';
-import {logger} from '@react-native-community/cli-tools';
+import {logger, CLIError} from '@react-native-community/cli-tools';
 import promiseWaterfall from './promiseWaterfall';
 import commandStub from './commandStub';
 import promisify from './promisify';
@@ -83,7 +82,7 @@ function unlink(args: Array<string>, ctx: ConfigT) {
   const {[packageName]: dependency, ...otherDependencies} = ctx.dependencies;
 
   if (!dependency) {
-    throw new Error(dedent`
+    throw new CLIError(`
       Failed to unlink "${packageName}". It appears that the project is not linked yet.
     `);
   }
@@ -137,7 +136,9 @@ function unlink(args: Array<string>, ctx: ConfigT) {
     })
     .catch(err => {
       logger.error(
-        `It seems something went wrong while unlinking. Error:\n${err.message}`,
+        `It seems something went wrong while unlinking. CLIError:\n${
+          err.message
+        }`,
       );
       throw err;
     });
