+++ packages/cli/src/tools/config/resolveReactNativePath.js
@@ -2,7 +2,7 @@
  * @flow
  */
 import path from 'path';
-import dedent from 'dedent';
+import {CLIError} from '@react-native-community/cli-tools';
 
 /**
  * Finds path to React Native inside `node_modules` or throws
@@ -17,7 +17,7 @@ export default function resolveReactNativePath(root: string) {
       }),
     );
   } catch (_ignored) {
-    throw new Error(dedent`
+    throw new CLIError(`
       Unable to find React Native files. Make sure "react-native" module is installed
       in your project dependencies.
