+++ packages/platform-android/src/commands/index.js
@@ -1,7 +1,7 @@
 /**
  * @flow
  */
-import logAndroid from './logAndroid/logAndroid';
-import runAndroid from './runAndroid/runAndroid';
+import logAndroid from './logAndroid';
+import runAndroid from './runAndroid';
 
 export default [logAndroid, runAndroid];
