+++ packages/platform-android/src/commands/logAndroid/index.js
