+++ packages/platform-ios/src/commands/logIOS/index.js
