+++ packages/tools/package.json
@@ -4,6 +4,7 @@
   "main": "build/index.js",
   "dependencies": {
     "chalk": "^1.1.1",
+    "dedent": "^0.7.0",
     "lodash": "^4.17.5",
     "mime": "^1.3.4"
   }
