+++ packages/tools/src/index.js
@@ -3,3 +3,5 @@
  */
 export {default as logger} from './logger';
 export {default as groupFilesByType} from './groupFilesByType';
+
+export * from './errors';
